nmrc poc

+++ FintrakBanking.Entities/Models/TBL_LOAN_APPLICATION_DETAIL.cs
@@ -169,7 +169,7 @@ namespace FintrakBanking.Entities.Models
         public bool? ISMORATORIUM { get; set; }
         public decimal? APPROVEDLINELIMIT { get; set; }
         public int? APPROVEDTRADECYCLEID { get; set; }
-        public int? PROPERTYTYPEID { get; set; }
+        public short? PROPERTYTYPEID { get; set; }
         public string PROPERTYTITLE { get; set; }
         public decimal? PROPERTYPRICE { get; set; }
         public decimal? DOWNPAYMENT { get; set; }
